requirements.txt
streamlit>=1.30.0
pandas>=2.0
numpy>=1.24
plotly>=5.18
streamlit-aggrid>=0.3.4
scikit-learn>=1.3
fpdf2>=2.7.0
